+++ chat_test.go
@@ -1,4 +1,1 @@
-//go:build !test
-// +build !test
-
 package groq
