+++ client_test.go
@@ -1,6 +1,3 @@
-//go:build !test
-// +build !test
-
 package groq
 
 import (
